% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datamining_ge20200306.R
\name{ge.plot.boxplot}
\alias{ge.plot.boxplot}
\title{ge.plot.boxplot}
\usage{
ge.plot.boxplot(data, x, y, type, filename)
}
\arguments{
\item{filename}{}
}
\description{
ge.plot.boxplot
}
