% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datamining_ge20200306.R
\name{ge.plot.pool.correlation}
\alias{ge.plot.pool.correlation}
\title{ge.plot.pool.correlation}
\usage{
ge.plot.pool.correlation(data, name = "bio_cor", method = "circle")
}
\arguments{
\item{method}{}
}
\description{
ge.plot.pool.correlation
}
