% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datamining_ge20200306.R
\name{ge.plot.techrep.correlation}
\alias{ge.plot.techrep.correlation}
\title{ge.plot.techrep.correlation}
\usage{
ge.plot.techrep.correlation(cor1, cor2, name = "pearson_correlation")
}
\arguments{
\item{name}{}
}
\description{
ge.plot.techrep.correlation
}
