% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datamining_ge20200306.R
\name{ge.plot.valcano}
\alias{ge.plot.valcano}
\title{ge.plot.valcano}
\usage{
ge.plot.valcano(data, title, fd = 1, pvalue = 0.05)
}
\arguments{
\item{pvalue}{}
}
\description{
ge.plot.valcano
}
