% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datamining_ge20200306.R
\name{ge.remove.techrep}
\alias{ge.remove.techrep}
\title{ge.remove.techrep}
\usage{
ge.remove.techrep(data, pattern = "_repB", method = "mean")
}
\arguments{
\item{method}{}
}
\description{
ge.remove.techrep
}
