% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datamining_ge20200306.R
\name{ge.writetable}
\alias{ge.writetable}
\title{ge.writetable}
\usage{
ge.writetable(data, filename, sep = "\\t", col.names = T,
  row.names = T, quote = F)
}
\arguments{
\item{quote}{}
}
\description{
ge.writetable
}
